Push down I2C logic into bsp-specific code
