+++ task/thermal/src/bsp/gimlet_bcdef.rs
@@ -6,9 +6,8 @@
 
 use crate::{
     control::{
-        ChannelType, Device, FanPresence, FanReading, InputChannel,
-        InputChannelMetadata, InputReadingOutcome, InputStatus, Max31790State,
-        PidConfig, TemperatureSensor,
+        ChannelType, FanPresence, FanReading, InputReadingOutcome, InputStatus,
+        PidConfig,
     },
     i2c_config::{devices, sensors},
 };
@@ -21,6 +20,17 @@ use userlib::{
     TaskId, UnwrapLite, task_slot,
     units::{Celsius, PWMDuty},
 };
+
+// This BSP uses i2c temperature inputs
+#[path = "./common/i2c_temp_input.rs"]
+mod i2c_temp_input;
+use i2c_temp_input::{
+    Device, InputChannel, InputChannelMetadata, TemperatureSensor,
+};
+
+#[path = "./common/max31790.rs"]
+mod max31790;
+use max31790::Max31790State;
 
 task_slot!(SEQ, gimlet_seq);
 
@@ -166,39 +176,7 @@ impl crate::control::BspInterface for Bsp {
     }
 
     fn read_fan_rpms(&mut self) -> impl Iterator<Item = FanReading> {
-        // Try to initialize the fan controller once at the start of the loop
-        let mut fctrl =
-            self.fctrl.try_initialize().map_err(SensorReadError::from);
-
-        // TODO: Maybe there's a way to make this a method on Fan that we can
-        // call, kind of like InputStatus?
-        self.fans.iter_mut().map(move |f| {
-            // If initialization failed, then we short circuit to return that
-            // original error, copied for each fan we're going to report.
-            let fctrl = fctrl.as_mut().map_err(|e| *e);
-
-            // If it was a success, attempt to read the RPMs, and either report
-            // that success or that error for each fan rpm.
-            let res = fctrl.and_then(|fc| {
-                fc.fan_rpm(f.bsp_data).map_err(SensorReadError::I2cError)
-            });
-            match res {
-                Ok(rpm) => {
-                    f.last_reading = Some(rpm);
-                    FanReading::PresentSuccess {
-                        rpm,
-                        sensor_id: f.rpm_sensor_id,
-                    }
-                }
-                Err(error) => {
-                    f.last_reading = None;
-                    FanReading::PresentError {
-                        error,
-                        sensor_id: f.rpm_sensor_id,
-                    }
-                }
-            }
-        })
+        self.fctrl.read_fan_rpms(self.fans)
     }
 
     fn read_misc_sensors(
